requirements.txt
Flask>=3.0.0
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0
requests>=2.31.0
joblib>=1.3.2
matplotlib>=3.8.0
seaborn>=0.13.0
Werkzeug>=3.0.0
python-dotenv>=1.0.0
